move resign into main components, git auth in dev and test only
